fix next and prev buttons
